Add support for custom status line information
If the environment variable FBPAD_STATUS is set,
it will be treated as the path to a file.
The first line of the file is read and displayed
on the right side of the status line and the status
line is made toggle-able. Note for long lines, only
the first 32 characters are printed. This is currenly
hard-coded, but should probably be configurable or computed.

If the FBPAD_STATUS variable is not set,
the bar will behave as usual.

Fbpad needs to be notified to reflect any chances to
the status file. When fbpad receives SIGARLM,
it will reread the status file and update the status line.

If the environment variable STATUS_PID is set,
fbpad will exec into `kill' when it is finished to
send SIGTERM to the pid indicated by STATUS_PID.

These mechanisms are meant to help with scripts that
update the status file information in the background.