Authentication now works differently:
Instead of using a password to authentiate via a local
ssh server, check the password against a private key
in the user's .ssh directory.
No connection to any ssh server is used. We only check
that importing the private key is successful.
Users should set an appropriate passphrase for the keyfile
they want to use using ssh-keygen or such.
This allows users to set a different password for unlocking
fbpad than their login password if they desire.
The name of the keyfile is configured in conf.h, but
fbpad will only look for it in ~/.ssh